Adding our 404 page
Put `not-found.html` inside of pages. Wasn't quite sure exactly how to
set up the block correctly for this instance and so need some guidance
on how to rework everything. As far as how to do the styling that
shouldn't matter too much. There is like one extra div
`mystery-location` just incase we wanted to do anything with it in the
future.